requirements.txt
numpy
pyomo
jmetalpy
#pandas
#plotly
#scipy
#jmetalpy
#pyscipopt
#gurobipy